feat(browser-clipboard): new @kolu/browser-clipboard — 6th framework this loop
Cross-cutting infrastructure for non-secure-context clipboard
writes. `navigator.clipboard === undefined` on any host that isn't
`https://…`, `localhost`, or `127.0.0.1`; the `.writeText` call
throws `TypeError` with no permission prompt to recover with. This
hits anyone running Kolu over a LAN address, machine hostname, or
Tailscale IP — exactly the dogfood path.

Fallback: `document.execCommand("copy")` against a synthetic
off-screen `<textarea>`. Formally deprecated but at
caniuse 100/100 with no removal timeline — the Clipboard API has
no equivalent fallback for non-secure contexts.

Two entry points so consumers without xterm don't pay the
`@xterm/addon-clipboard` peer-dep cost:

- `./` — `writeTextToClipboard(text)`.
- `./xterm` — `SafeClipboardProvider` implementing xterm's
  `IClipboardProvider` (OSC 52 writes survive plain HTTP; reads
  return empty when `navigator.clipboard.readText` is unavailable).

8 client consumers swap from `ui/clipboard` to
`@kolu/browser-clipboard` (Terminal.tsx imports both entry points).

161/161 client unit tests pass.
