+++ README.md
@@ -201,6 +201,7 @@ pnpm monorepo:
 | `packages/solid-canvas-viewport/`    | Solid-native pan/zoom 2D canvas viewport. Wheel-toward-cursor zoom, click/middle/shift-drag pan, animated `centerOnRect` with `prefers-reduced-motion`, snap-to-grid. Decomposed by volatility: gesture input (`gestures.ts`), transform math (`transforms.ts` + `animatedPan.ts`), CSS output (`coordinates.ts` — `tileTransformCSS`). `useCanvasViewport()` orchestrates the three behind one `CanvasViewport` interface. Single consumer (`client/src/canvas/`) |
 | `packages/solid-recorder/`           | Solid-native browser tab + mic (+ optional webcam) recording. `useRecorder()` owns the idle→setup→recording→paused state machine, elapsed clock, device enumeration, and the WebM duration patch. `mic` / `webcam` are singleton domain state. `LevelMeter` + `WebcamOverlay` are the framework-supplied visual pieces. Notification surface (`onError` / `onSuccess` / `onWarning`) is injectable via `configureRecorderNotifications` — Kolu wires `solid-sonner` through it at app init. Single consumer (`client/src/recorder/`) |
 | `packages/solid-anchored-popover/`   | Minimal SolidJS popover-positioning primitive. `useAnchoredPopover({ triggerRef, open, onDismiss, anchor })` returns `{ panelRef, panelStyle }` and owns viewport-clamped positioning + outside-click + Escape dismiss. Six in-tree Kolu consumers (option menu, settings popover, record popover, mode-chip picker, activity-window chip, PR-unavailable tooltip) — the highest-reuse primitive Kolu had |
+| `packages/browser-clipboard/`        | Clipboard write that survives non-secure contexts. `navigator.clipboard` is exposed only in `https://…` / `localhost` / `127.0.0.1`; plain `http://` to a LAN address gets `navigator.clipboard === undefined`. `writeTextToClipboard(text)` tries the modern API and falls back to `document.execCommand("copy")` against a synthetic off-screen `<textarea>` ([caniuse 100/100](https://caniuse.com/mdn-api_document_execcommand_copy)). Separate `./xterm` entry point exports `SafeClipboardProvider` implementing xterm.js's `IClipboardProvider` so OSC 52 writes survive plain HTTP |
 | `packages/artifact-sdk/`             | Self-contained comments-on-files toolkit. Three exports: `./core` (pure W3C TextQuoteSelector functions used by both runtimes), `./client` (parent-side iframe ↔ parent bridge + core re-exports), `./server` (one-line `mountArtifactSdk(app, ...)` that wires the SDK bundle route and an HTML-decoration middleware — esbuild bundles the in-iframe script at server startup, hash-keyed for cache busting). The host server never imports HTML rewriting logic |
 | `packages/server/`                   | [Hono](https://hono.dev/) + `kolu-pty` for terminal lifecycle                                                                                                          |
 | `packages/integrations/pty/`         | Generic PTY primitives — [node-pty](https://github.com/microsoft/node-pty) + [@xterm/headless](https://www.npmjs.com/package/@xterm/headless) wrapped behind a `PtyHandle` with OSC-driven cwd/title/preexec callbacks; only kolu-* dep is `kolu-shared` (Logger type) |
