+++ docs/architecture-ralph-report.md
@@ -135,7 +135,8 @@ The repo has clearly internalized the extraction discipline:
 ---
 
 |    11 | `useAnchoredPopover.ts` (135 LOC, 6 in-tree consumers) | extract `@kolu/solid-anchored-popover` — move the file bit-for-bit (it's already at the right altitude), update 6 client consumers to import from the new package | 5th `@kolu/*` framework this loop; first one extracted with a true *in-tree* high-reuse justification (not just the volatility-axis bar) | new package `+4` files; client/ui `−1` file | `8647a71c` | Out of /lowy on `server/terminalBackend/local.ts`: no actions warranted there (the meta/* consolidation is correct). Pivoted to the highest-in-tree-reuse SolidJS primitive: 6 consumers of `useAnchoredPopover` (option menu, settings popover, record popover, mode-chip picker, activity-window chip, PR-unavailable tooltip). The hook only depends on `solid-js` + `@solid-primitives/event-listener` — zero Kolu coupling. README explicitly addresses "why not Corvu / Floating UI" (this hook stays close to imperative `getBoundingClientRect` so consumers own the DOM + portal target). 161/161 client unit tests pass. |
-|    12 | naming convention: `terminal-themes` was unscoped despite being externalization-ready (581 LOC of curated terminal color schemes + perceptual-distance picker) | rename to `@kolu/terminal-themes`; update 7 client import sites (+ `./color` subpath); update package.json `dependencies`; refresh README header | aligns naming with external-publish intent (joining `@kolu/surface`, `@kolu/solid-pierre`, the 5 framework packages added this loop, and `@kolu/artifact-sdk`); no behaviour change | 7 import-site edits | (cycle 12) | The package was already at the right altitude — pure data + perceptual-distance picker, zero Kolu app coupling. The unscoped name was a remnant from before the `@kolu/*` convention crystallized. README rewrites the example to match. Sets the convention for the remaining unscoped packages (`anyagent`, `nonempty`, `memorable-names`) which can follow similarly in future cycles. 161/161 client unit tests pass. |
+|    12 | naming convention: `terminal-themes` was unscoped despite being externalization-ready (581 LOC of curated terminal color schemes + perceptual-distance picker) | rename to `@kolu/terminal-themes`; update 7 client import sites (+ `./color` subpath); update package.json `dependencies`; refresh README header | aligns naming with external-publish intent (joining `@kolu/surface`, `@kolu/solid-pierre`, the 5 framework packages added this loop, and `@kolu/artifact-sdk`); no behaviour change | 7 import-site edits | `08b04ff7` | The package was already at the right altitude — pure data + perceptual-distance picker, zero Kolu app coupling. The unscoped name was a remnant from before the `@kolu/*` convention crystallized. README rewrites the example to match. Sets the convention for the remaining unscoped packages (`anyagent`, `nonempty`, `memorable-names`) which can follow similarly in future cycles. 161/161 client unit tests pass. |
+|    13 | `client/src/ui/clipboard.ts` (111 LOC, 6 consumers) — cross-cutting infrastructure for non-secure-context clipboard writes | extract `@kolu/browser-clipboard` with `./` entry (`writeTextToClipboard`) and `./xterm` entry (`SafeClipboardProvider` impl of xterm's `IClipboardProvider`); 8 client consumers updated | 6th `@kolu/*` framework this loop; documents the LAN-HTTP/Tailscale-IP volatility axis that any browser app over plain HTTP hits | new package `+5` files; client/ui `−1` file | (cycle 13) | Real cross-cutting concern: `navigator.clipboard === undefined` for any host that isn't `https://…`, `localhost`, or `127.0.0.1`. The `execCommand("copy")` + synthetic `<textarea>` fallback is formally deprecated but at [caniuse 100/100](https://caniuse.com/mdn-api_document_execcommand_copy) with no removal timeline. Two entry points so consumers without xterm don't pay the `@xterm/addon-clipboard` peer-dep cost. 161/161 client unit tests pass. |
 
 ---
