+++ packages/browser-clipboard/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@kolu/browser-clipboard",
+  "version": "0.1.0",
+  "private": true,
+  "license": "AGPL-3.0-or-later",
+  "type": "module",
+  "sideEffects": false,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./xterm": "./src/xterm.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.8.0"
+  },
+  "dependencies": {
+    "@xterm/addon-clipboard": "^0.2.0"
+  }
+}
