+++ packages/client/src/ui/pierreAdapters.ts
@@ -12,7 +12,7 @@ import type {
 } from "@kolu/solid-pierre";
 import type { GitChangeStatus } from "kolu-git/schemas";
 import { toast } from "solid-sonner";
-import { writeTextToClipboard } from "./clipboard";
+import { writeTextToClipboard } from "@kolu/browser-clipboard";
 
 const GIT_STATUS_WORD: Record<GitChangeStatus, GitStatusEntry["status"]> = {
   M: "modified",
